+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to the "unsaved-files-vscode" extension will be documented i
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## 1.3.0 - 2020-11-??
+
+### Added
+
+- `openInGithubDesktop.diagnosticWarning` setting
+- `openInGithubDesktop.unsavedWarning` setting
+
 ## 1.2.1 - 2020-10-22
 
 ### Fixed